requirements.txt
certifi
Pillow
pyserial
wincertstore
ttkbootstrap
matplotlib
pyinstaller